move CurrencyBeaconService to a separate file
