+++ src/routes/index.ts
@@ -1,12 +1,12 @@
 import express from 'express';
 
 import * as controllers from '../controllers';
-import { CurrencyBeaconService } from '../services/RateSourceService';
+import { CurrencyBeaconService } from '../services/rateSource/CurrencyBeaconService';
 
 const router = express.Router();
 
 const rateController = new controllers.RateController(
-  new CurrencyBeaconService(process.env.CURRENCY_BEACON_API_KEY ?? ''),
+  new CurrencyBeaconService(process.env.RATE_SOURCE_API_KEY ?? ''),
 );
 const subscribeController = new controllers.SubscribeController();
 const unsubscribeController = new controllers.UnsubscribeController();
