+++ src/services/rateSource/RateSourceService.ts
@@ -1,0 +1,5 @@
+export abstract class RateSourceService {
+  protected constructor(protected readonly url: string) {}
+
+  abstract retrieve(): Promise<number | undefined>;
+}
